pyproject.toml
[tool.poetry]
name = "drfapi"
version = "0.1.0"
description = ""
authors = ["joudi awameh <awamehj@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
docker = "^4.4.1"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
